# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is a sample Petstore server.  You can find\n*out more about Swagger at\n*[http://swagger.io](http://swagger.io) or on\n*[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/family/child/budi": {
            "get": {
                "tags": [
                    "Family"
                ],
                "summary": "Find child budi",
                "description": "look for data on the child from Budi",
                "operationId": "child",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/family/grandchild/budi": {
            "get": {
                "tags": [
                    "Family"
                ],
                "summary": "Find child budi",
                "description": "look for data on the child from Budi",
                "operationId": "grandchild",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender filter",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "laki-laki",
                                "perempuan"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/family/cousin/hani": {
            "get": {
                "tags": [
                    "Family"
                ],
                "summary": "Find child budi",
                "description": "look for data on the child from Budi",
                "operationId": "cousin",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender filter",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "laki-laki",
                                "perempuan"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/family": {
            "get": {
                "tags": [
                    "Family"
                ],
                "summary": "Find child budi",
                "description": "look for data on the child from Budi",
                "operationId": "families",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender filter",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "laki-laki",
                                "perempuan"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "deprecated": false
            }
        }
    }
}